AUT-5553: Send update request to account data api when user has successfully verified with passkey
To do: handle failures, get user profile in safe way
This necessitated changing the type of the sign count to a long since this is what we get from the assertion result. Updates on the account data api to come in subsequent commit.

+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/services/passkeys/PasskeysService.java
@@ -117,7 +117,7 @@ public class PasskeysService {
             String publicSubjectId,
             String sessionId,
             String passkeyIdentifier,
-            int signCount,
+            long signCount,
             Clock clock) {
         var accountDataBaseUri = configurationService.getAccountDataURI();
         var updatePasskeysRequestUri =
